Simplify layouting by utilizing CSS grid areas
No longer unnecessarily define width, float, etc. by utilizing grid-template-areas. Rearrange the sections more clearly and define dividers for better content padding, define the card contents as flex and have the horizontal and vertical alignments be set to center. Also set the timing function for pressing the cards to ease-out and adjust the design of the layout slightly.

Signed-off-by: ederekun <sedrickvince@gmail.com>