jQuery: Initial integration
This commit integrates the usage of jQuery into the website.

For this use case, instead of having to scroll through the website,
jQuery is used to update main section-container in index.html based
on HTML data split accordingly to sections under the section folder.
Additionally, we use jQuery to animate this process of updating the
website's content by using the built in show/hide functionality.

Signed-off-by: ederekun <sedrickvince@gmail.com>