jQuery: Adjust CSS config and add nav-selected
With the new jQuery way of updating the website, let's adjust certain
classes in style.css as we had formatted the website before according
to how it appears as we scroll though it.

In addition to this, add a new class that we remove and add every time
we select a section. This will allow the nav-buttons to retain the
inverted colors as they get selected.

Lastly, let's also fix the indentions within the file while
we're at it and remove the header at projects.html as it becomes
redundant with the new nav-selected class added.

Signed-off-by: ederekun <sedrickvince@gmail.com>