sections: Update based on latest information
Update all sections accordingly with what I have so far. In addition
to this, replace nationality with relevant experiences as to have better
focus and relevance. Also add descriptions to each personal projects as to
describe the purpose of their creation.

Alongside this update, implement a better layout for the about section to
avoid visual bugs when the viewport adjusts.

Signed-off-by: ederekun <sedrickvince@gmail.com>