style.css: Modify navbar positioning and slight tweaks
Having the navbar fixed/sticky creates problems as the navbar overlaps
the section-container (now section-content). Instead of having it
being retained at the top, let's instead create an effect at which the
navbar gets "minimized" everytime we select a section; to "open" the navbar
again, we just simply scroll up.

In addition to this, fix the alignment of the profile photo.

Signed-off-by: ederekun <sedrickvince@gmail.com>