jQuery: Make the website responsive to history (popstate) navigations
Add an event listener to popstate for the website to update when the user
clicks the back/forward button. This allows us to simply invoke a pushstate
with the navigation buttons and also separate the URL key updates from the
section update function - making the updateSection's purpose solely for
updating the website content. Alongside this addition, also refactor the
whole script for better readability and avoid redundancy at certain
code paths.

Signed-off-by: ederekun <sedrickvince@gmail.com>